more work on GPS
now the image that the user uploads when setting up the his account will be used to mark his location. uploading an image is mandatory, no default images. TODO: 1) upload user location every 10 or so minutes. 2) resize the images the users choose on sign up to be 80x80 and then store them in firebase storage